backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
joblib==1.3.1
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
lightgbm==4.0.0
psycopg2-binary==2.9.7